Extended and improved QC filming
